Rename bomb package
